+++ apps/dashboard/src/components/primitives/button.tsx
@@ -59,7 +59,7 @@ export const buttonDeprecatedVariants = tv({
       },
       xs: {
         root: 'h-8 gap-2.5 rounded-lg px-2 text-label-xs',
-        icon: '',
+        icon: 'size-4',
       },
       '2xs': {
         root: 'h-7 gap-2.5 rounded-lg px-2 text-label-xs',
