+++ apps/dashboard/src/components/workflow-editor/test-workflow/test-workflow-tabs.tsx
@@ -13,12 +13,12 @@ import { buildRoute, ROUTES } from '@/utils/routes';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { type WorkflowTestDataResponseDto } from '@novu/shared';
 import { toast } from 'sonner';
-import { ButtonDeprecated } from '../../primitives/button-deprecated';
 import { Form } from '../../primitives/form/form';
 import { showToast } from '../../primitives/sonner-helpers';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '../../primitives/tabs';
 import { buildDynamicFormSchema, TestWorkflowFormType } from '../schema';
 import { TestWorkflowForm } from './test-workflow-form';
+import { Button } from '../../primitives/button';
 
 export const TestWorkflowTabs = ({ testData }: { testData?: WorkflowTestDataResponseDto }) => {
   const { environmentSlug = '', workflowSlug = '' } = useParams<{ environmentSlug: string; workflowSlug: string }>();
@@ -99,16 +99,16 @@ export const TestWorkflowTabs = ({ testData }: { testData?: WorkflowTestDataResp
                 </Link>
               </TabsTrigger>
               <div className="my-auto ml-auto flex items-center gap-2">
-                <ButtonDeprecated
+                <Button
                   type="submit"
                   variant="primary"
-                  size="sm"
-                  className="flex gap-1"
+                  size="xs"
+                  mode="gradient"
                   isLoading={isPending}
+                  leadingIcon={RiPlayCircleLine}
                 >
-                  <RiPlayCircleLine className="size-5" />
-                  <span>Test workflow</span>
-                </ButtonDeprecated>
+                  Test workflow
+                </Button>
               </div>
             </TabsList>
             <TabsContent value="trigger" className="mt-0 flex w-full flex-1 flex-col overflow-hidden" variant="regular">
